+++ test/tools/analyze_tool_test.dart
@@ -1,6 +1,4 @@
 @TestOn('vm')
-import 'dart:io';
-
 import 'package:args/args.dart';
 import 'package:args/command_runner.dart';
 import 'package:dart_dev/src/dart_dev_tool.dart';